Align test project settings with current Django versions (3.2/4.0)
The Django settings long use pathlib for operations related to directories and files.
The TEMPLATES[OPTIONS] dictionary has more context processors by default (from Django 1.8 on or earlier).
Django 4.0 now deprecates the use of the USE_L10N member in the Django settings.
The DEFAULT_AUTO_FIELD is new since Django 3.2 and issues a warning when not explicitly defined.