Speed up linting jobs, rely on defaults for Pip & Coverage
Pip is installed natively on Azure for running Python package installations.
It's in Microsoft's interest to keep the version sufficiently up-to-date.
Hence, we shouldn't need to bother about it.

For linting jobs we can tell Tox to not install our software as a package,
which is only an unnecessary overhead for static inspection.

We can use the latest version of Bandit when using the appropriate
configuration switch for ignoring irrelevant files and directories.

For coverage and unit test reports it makes sense to stick to the
default file names and locations, to avoid the need for explicit
configuration and to not confuse contributors and interested developers.